test(security): complete SEC-37 coverage vectors and strategy tests
Add deterministic known-answer vectors for v2 nonce/AAD at index 0, MaxUint64, and the terminal record, with terminal/data nonce noncollision assertions. Add boundary KAT bytes to ADR 0016. Add focused engine tests for AuthenticateChunkedTrailer covering password and mock-KeyManager paths, cancellation, wrong key, and DEK lifetime/error behavior. Add direct strategy-level tests for uploadPartCopyChunked and uploadPartCopyReencryptMPU covering success, range clamping, authenticated bounds, source/decrypt/upload errors, claim-store release, legacy append-store failure, AlreadyDone reservation, and cap exceeded paths. Add coverage matrix and version-aware range decrypt tests. Update existing tests for v2 compatibility and KeyManager context/zeroization observability.

- internal/crypto/sec37_test.go

- internal/crypto/sec37_coverage_matrix_test.go

- internal/crypto/range_decrypt_version_test.go

- internal/crypto/sec37_phase_a_test.go

- internal/crypto/chunked_test.go

- internal/crypto/range_decrypt_test.go

- internal/crypto/range_optimization_test.go

- internal/crypto/fuzz_test.go

- internal/crypto/engine_options_test.go

- internal/crypto/keymanager_cosmian_json_test.go

- internal/crypto/keymanager_openbao_test.go

- internal/api/sec37_test.go

- internal/api/upload_part_copy_strategy_test.go

- internal/api/coverage_gaps_test.go

- internal/api/encrypted_mpu_handler_chacha_test.go

- internal/api/encrypted_mpu_handler_test.go

- internal/audit/audit_test_helpers.go

- internal/audit/classifier_test.go

- internal/audit/inspect_test.go

Relates to docs/plans/V1.0-SEC-37-plan.md (Phases B-F).

+++ internal/crypto/chunked_test.go
@@ -762,7 +762,7 @@ func TestChunkedEngineDecrypt_LegacyObject(t *testing.T) {
 	manifest := &ChunkManifest{
 		Version:    1,
 		ChunkSize:  chunkSize,
-		ChunkCount: chunkCount,
+		ChunkCount: uint64(chunkCount),
 		BaseIV:     encodeBase64(baseIV),
 	}
 	manifestEncoded, err := encodeManifest(manifest)
